updated login.php, updated main.php
main.php - sessions now only start if a current session is not in use (once), PHPSESSID cookie is now unset when logout.php endpoint is called. (FUNCTIONAL)
login.php - now has an optional "authenticate" parameter which can be requested to when a webpage loads to check whether the session has expired. (FUNCTIONAL)